+++ server/controllers/uploadColController.js
@@ -3,9 +3,9 @@ import path from "path";
 import Column from '../models/Column.js';
 import { roleValidation } from "../services/roleValidation.js";
 
-export const handleUpload = async (req, res) => {
+export const handleUploadCol = async (req, res) => {
     
-    roleValidation(req, res, 'addReag')
+    if(!roleValidation(req, res, 'addColumn')) return;
     const handleIsURL = (str) =>  {
         const pattern = new RegExp('^(https?:\\/\\/)?'+ // protocol
           '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.)+[a-z]{2,}|'+ // domain name
@@ -21,23 +21,32 @@ export const handleUpload = async (req, res) => {
         const column = await Column.findOne({itemId}) 
 
         if(!column) return res.status(400).json({ message: "couldnt upload", clientMessage: "ошибка при загрузке файла паспорта"});
-        const fileName = req.body.fileName
-        // if (column.passport && handleIsURL(column.passport)){
-        if (column.passport && !handleIsURL(column.passport)){
-            const file = path.resolve('./docs/'+column.passport)
-            unlink(file, (err) => {
-            if (err) throw err;
-            console.log(file, ' was deleted');
-            });
+
+        const fileName = req.body.fileName;
+       
+        if (`${itemId} -- ${fileName}` === column.passport){
+            return res.sendStatus(200)
         }
         
-        const name = `${itemId}--${fileName}`
+        if (column.passport && !handleIsURL(column.passport)){
+            const file = path.resolve('./docs/'+column.passport);
+            unlink(file, (err) => {
+                if (err) {
+                    console.log(err);
+                } else {
+                    console.log(file, ' was deleted'); 
+                }             
+            });
+        }
+
+        const name = `${itemId} -- ${fileName}`
         column.passport = name
         column.save();
         res.json({message: 'uploaded'})
 
     } catch (error) {
         res.status(500)
+        console.log(error)
         .json({
             message: 'server side error', 
             clientMessage: 'Ошибка сервера при загрузке пасспорта',
